RefSW: Refactor
