+++ libswirl/rend/soft/refsw_pixel.h
@@ -1,0 +1,9 @@
+#pragma once
+/*
+	This file is part of libswirl
+*/
+#include "license/bsd"
+
+#include "refsw.h"
+
+PixelPipeline* Create_RefPixelPipeline();
